Added an option (spark.closure.serializer) to specify the serializer for closures. This enables using Kryo as the closure serializer.
